# benchmarks: add metrics benchmark